% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-gallery.R
\name{gallery}
\alias{gallery}
\title{Create an Image Gallery Widget}
\usage{
gallery(values, options = list(), width = NULL, height = NULL)
}
\arguments{
\item{values}{File or URL paths to images that are to be displayed.}

\item{options}{A list of initialization options.}

\item{width}{The width of the input container, e.g., \code{'100\%'}; see
\link{validateCssUnit}.}

\item{height}{The height of each input image, e.g., \code{'200px'}; see
\link{validateCssUnit}.}
}
\value{
An image gallery widget that can be added to a UI definition.
}
\description{
Create an image gallery widget with pagination based on file or
URL paths to images.
}
